Minor: move datasource statistics code into its own module
